requirements.txt
# requirements for django-datafile

django>=1.7
django-model-utils>=2.0
django-taggit>=0.12
django-taggit-templatetags2
django-bootstrap3

-e git+https://github.com:pmeier82/spikeval#egg=spikeval
